+++ progetto-fermi-repository-adapter/src/main/java/it/previnet/progettofermi/repository/adapter/AbstractRepositoryImpl.java
@@ -22,12 +22,12 @@ public abstract class AbstractRepositoryImpl<T> implements AbstractRepository<T>
     }
 
     @Override
-    public T findByToken(Long token) {
+    public T findByToken(Integer token) {
         return getEntityManager().find(entityType, token);
     }
 
     @Override
-    public T findReference(Long token) {
+    public T findReference(Integer token) {
         return getEntityManager().getReference(entityType, token);
     }
